CLI login prompt that prompts user to try until 3 failed attempts
