+++ newbank/server/src/main/Customer.java
@@ -1,4 +1,4 @@
-package com.server;
+package newbank.server.src.main;
 
 import java.util.ArrayList;
