+++ newbank/server/src/main/CustomerID.java
@@ -1,4 +1,4 @@
-package com.server;
+package newbank.server.src.main;
 // no explicit boilerplate constructor needed if a record is used to store the record
 
 public class CustomerID {
