+++ newbank/server/src/main/Password.java
@@ -1,4 +1,4 @@
-package com.server;
+package newbank.server.src.main;
 
 public class Password {
     private String password;
