+++ newbank/server/src/main/resources/META-INF/MANIFEST.MF
@@ -1,1 +1,1 @@
-Main-Class: com.server.NewBankServer
+Main-Class: newbank.server.src.main.NewBankServer
