+++ newbank/server/src/test/java/com/server/MoveTest.java
@@ -1,4 +1,4 @@
-package com.server;
+package newbank.server.src.test.java.com.server;
 
 import org.junit.Test;
 import org.junit.Before;
